bot/requirements.txt
aiormq==6.7.4
anyio==3.6.2
asttokens==2.0.5
asyncpg==0.27.0
autoflake==2.0.2
black==23.3.0
carrot-rpc==0.3
certifi==2022.12.7
cfgv==3.3.1
channel-box==0.4.0
classify-imports==4.2.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
executing==0.8.1
fastapi==0.95.1
filelock==3.11.0
flake8==6.0.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
orjson==3.8.9
packaging==23.0
pamqp==3.2.1
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.12.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyupgrade==3.3.1
PyYAML==6.0
reorder-python-imports==3.9.0
shortuuid==1.0.11
simple-print==1.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
termcolor==1.1.0
tokenize-rt==5.0.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.21.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.8.2
yesqa==1.4.0
